DraftClone/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.21
httpx==0.27.2
PyJWT==2.9.0
aiogram==3.13.1
python-dotenv==1.0.1
